FAC-572: Refactor - Transfer 'NOTIFICATION_TITLE's from 'ErrorMessageKey' to 'MessageKey'
