+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/service/assessment/notification/CreateAssessmentNotificationCreator.java
@@ -15,7 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-import static org.flickit.assessment.core.common.ErrorMessageKey.NOTIFICATION_TITLE_CREATE_ASSESSMENT;
+import static org.flickit.assessment.core.common.MessageKey.NOTIFICATION_TITLE_CREATE_ASSESSMENT;
 
 @Slf4j
 @Component
