+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/common/MessageKey.java
@@ -12,4 +12,10 @@ public class MessageKey {
     public static final String ASSESSMENT_DEFAULT_INSIGHT_DEFAULT_INCOMPLETE = "assessment-default-insight.incomplete";
 
     public static final String SUBJECT_DEFAULT_INSIGHT = "subject-default-insight";
+
+    public static final String NOTIFICATION_TITLE_GRANT_ASSESSMENT_USER_ROLE = "notification-title.grant-assessment-user-role";
+    public static final String NOTIFICATION_TITLE_CREATE_ASSESSMENT = "notification-title.create-assessment";
+    public static final String NOTIFICATION_TITLE_COMPLETE_ASSESSMENT = "notification-title.completed-assessment";
+    public static final String NOTIFICATION_TITLE_ACCEPT_ASSESSMENT_INVITATION = "notification-title.accept-assessment-invitation";
+
 }
